订单状态定时处理

+++ take-out-pojo/src/main/java/com/player32611/entity/Orders.java
@@ -84,7 +84,7 @@ public class Orders implements Serializable {
     private LocalDateTime deliveryTime;
 
     /** 打包数量 (INT) */
-    private Integer packAmount;
+    private BigDecimal packAmount;
 
     /** 餐具数量 (INT) */
     private Integer tablewareNumber;
